locatePreviewPlacer based on window not screen
locatePreviewPlacerView() should not compare a relative window
coordinate to a fixed screen display resolution metric when
determining if the Keyboard is in a transient state.

While it's reasonable for mPreivewPlacerView's actual drawing
location to be based on a relative coordinate within the window,
determining whether or not the keyboard is settled should be
compared based on its absolute position within the display.

The key preview may not be displayed when relying on a relative
window coordinate if the window offset in the display is small
enough, even if there's plenty of space available in the window
and on the display.

Change-Id: I32618500964789422f8d93996cf0c4cd08b5b104
Signed-off-by: Kyle Russell <bkylerussell@gmail.com>